requirements.txt
Django==2.0.5
djangorestframework==3.8.2
django-environ==0.4.4
dj-database-url==0.5.0
psycopg2-binary==2.7.4